CRE-483 - Fixes lint
