+++ system-tests/lib/cre/environment/jobs.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/smartcontractkit/chainlink-testing-framework/framework/components/blockchain"
 	"github.com/smartcontractkit/chainlink-testing-framework/framework/components/jd"
-	ns "github.com/smartcontractkit/chainlink-testing-framework/framework/components/simple_node_set"
 
 	"github.com/smartcontractkit/chainlink/system-tests/lib/cre/crib"
 	cretypes "github.com/smartcontractkit/chainlink/system-tests/lib/cre/types"
@@ -54,54 +53,6 @@ func StartJD(lggr zerolog.Logger, nixShell *nix.Shell, jdInput jd.Input, infraTy
 	lggr.Info().Msgf("Job Distributor started in %.2f seconds", time.Since(startTime).Seconds())
 
 	return jdOutput, nil
-}
-
-func StartDONs(
-	lggr zerolog.Logger,
-	nixShell *nix.Shell,
-	topology *cretypes.Topology,
-	infraType libtypes.InfraType,
-	registryChainBlockchainOutput *blockchain.Output,
-	capabilitiesAwareNodeSets []*cretypes.CapabilitiesAwareNodeSet,
-) ([]*cretypes.WrappedNodeOutput, error) {
-	startTime := time.Now()
-	lggr.Info().Msgf("Starting %d DONs", len(capabilitiesAwareNodeSets))
-
-	if infraType == libtypes.CRIB {
-		lggr.Info().Msg("Saving node configs and secret overrides")
-		deployCribDonsInput := &cretypes.DeployCribDonsInput{
-			Topology:       topology,
-			NodeSetInputs:  capabilitiesAwareNodeSets,
-			NixShell:       nixShell,
-			CribConfigsDir: cribConfigsDir,
-		}
-
-		var devspaceErr error
-		capabilitiesAwareNodeSets, devspaceErr = crib.DeployDons(deployCribDonsInput)
-		if devspaceErr != nil {
-			return nil, pkgerrors.Wrap(devspaceErr, "failed to deploy Dons with devspace")
-		}
-	}
-
-	nodeSetOutput := make([]*cretypes.WrappedNodeOutput, 0, len(capabilitiesAwareNodeSets))
-
-	// TODO we could parallelize this as well in the future, but for single DON env this doesn't matter
-	for _, nodeSetInput := range capabilitiesAwareNodeSets {
-		nodeset, nodesetErr := ns.NewSharedDBNodeSet(nodeSetInput.Input, registryChainBlockchainOutput)
-		if nodesetErr != nil {
-			return nil, pkgerrors.Wrapf(nodesetErr, "failed to create node set named %s", nodeSetInput.Name)
-		}
-
-		nodeSetOutput = append(nodeSetOutput, &cretypes.WrappedNodeOutput{
-			Output:       nodeset,
-			NodeSetName:  nodeSetInput.Name,
-			Capabilities: nodeSetInput.Capabilities,
-		})
-	}
-
-	lggr.Info().Msgf("DONs started in %.2f seconds", time.Since(startTime).Seconds())
-
-	return nodeSetOutput, nil
 }
 
 func SetupJobs(
